Add validation to application for bad request

+++ Application/Activities/Delete.cs
@@ -1,3 +1,4 @@
+using Application.Core;
 using MediatR;
 using Persistence;
 
@@ -5,12 +6,12 @@ namespace Application.Activities
 {
     public class Delete
     {
-        public class Command : IRequest
+        public class Command : IRequest<Result<Unit>>
         {
             public Guid ActivityId { get; set; }
         }
 
-        public class Handler : IRequestHandler<Command>
+        public class Handler : IRequestHandler<Command, Result<Unit>>
         {
             private readonly DataContext _context;
             public Handler(DataContext context)
@@ -18,13 +19,15 @@ namespace Application.Activities
                 _context = context;
             }
 
-            public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
                 var activity = await _context.Activities.FindAsync(request.ActivityId);
-                _context.Activities.Remove(activity);
-                await _context.SaveChangesAsync();
+                if (activity == null) return null;
 
-                return Unit.Value;
+                _context.Activities.Remove(activity);
+                var result = await _context.SaveChangesAsync() > 0;
+
+                return result ? Result<Unit>.Success(Unit.Value) : Result<Unit>.Failure("Failed to delete Activity");
             }
         }
     }
